Create SECURITY.md
The URL you provided points to the "New file" page in the `master` branch of the `vkb40/ethereal` repository. Since this is a page for creating a new file and not an existing file, there is no content to summarize. Could you clarify or provide the path to an existing file you'd like summarized?