Create SECURITY.md
The URL you provided points to the creation of a new file in the `master` branch of the `wealdtech/ethereal` repository. Since the file doesn't exist yet and no specific file is referenced, there is nothing to summarize. Let me know if you'd like assistance creating or editing a new file!